added some new icons
