+++ src/components/Icons/MicroscopeIcon.tsx
@@ -1,0 +1,19 @@
+function MicroscopeIcon() {
+  return (
+    <svg xmlns='http://www.w3.org/2000/svg' width='42' height='42' viewBox='0 0 42 42' fill='none'>
+      <path
+        d='M12.25 34.375C12.25 33.2704 13.1454 32.375 14.25 32.375H27.75C28.8546 32.375 29.75 33.2704 29.75 34.375V35.375C29.75 35.6511 29.5261 35.875 29.25 35.875H12.75C12.4739 35.875 12.25 35.6511 12.25 35.375V34.375Z'
+        stroke='#22202A'
+        stroke-width='1.5'
+      />
+      <path
+        d='M15.6598 26.25C15.6598 26.25 18.0451 28.0379 19.4901 28.3352C20.7777 28.6001 22.1123 28.5215 23.3599 28.1071C24.6074 27.6927 25.7239 26.9573 26.5971 25.9746C27.4704 24.992 28.0696 23.7968 28.3345 22.5093C28.5994 21.2217 28.5207 19.887 28.1064 18.6395C27.9079 18.0419 27.6357 17.4743 27.298 16.9492L27.1243 16.625M11.9062 26.25C11.9062 26.25 13.1802 28.0984 14.0245 28.8487C15.4022 30.0731 17.0779 30.9132 18.8833 31.2846C20.6886 31.6561 22.5599 31.5457 24.309 30.9648C26.0582 30.3838 27.6236 29.3526 28.848 27.9749C30.0723 26.5971 30.9124 24.9214 31.2839 23.1161C31.6553 21.3108 31.545 19.4395 30.964 17.6903C30.6592 16.7724 30.2303 15.9052 29.6925 15.1112L29.5636 14.9717M27.1243 16.625C28.2308 16.625 29.1774 15.9404 29.5636 14.9717M27.1243 16.625C26.1997 16.625 25.3867 16.1469 24.9191 15.4245M29.5636 14.9717C29.6834 14.6711 29.7493 14.3433 29.7493 14C29.7493 13.3237 29.4936 12.7072 29.0735 12.2418M29.8889 8.92162L29.7493 9.03025M29.8889 8.92162C30.2265 9.36227 30.807 9.37083 31.2159 9.04125C31.8263 8.54939 31.8874 7.67225 31.4222 7.06496L31.0018 6.51623C30.5366 5.90894 29.6738 5.73961 29.04 6.20086C28.6185 6.50756 28.4784 7.0804 28.81 7.51326M29.8889 8.92162L29.3495 8.21744L28.81 7.51326M28.81 7.51326L28.6726 7.62475M28.6726 7.62475L29.7493 9.03025M28.6726 7.62475C28.039 6.79772 27.7222 6.38421 27.2903 6.25366C27.1454 6.20987 26.9944 6.18986 26.8431 6.19443C26.392 6.20803 25.9785 6.52482 25.1515 7.15839L16.4135 13.8524C16.1786 14.0324 16.0611 14.1224 15.9766 14.2152C15.5834 14.6473 15.4998 15.2782 15.767 15.7978C15.8244 15.9093 15.9144 16.0268 16.0944 16.2618M29.7493 9.03025C30.168 9.57677 30.3773 9.85003 30.4471 10.1359C30.5302 10.4763 30.4826 10.8355 30.3138 11.1425C30.172 11.4003 29.8987 11.6096 29.3522 12.0283L29.0735 12.2418M16.0944 16.2618L18.2763 19.1099M16.0944 16.2618L14.3031 17.6353C13.883 17.9571 13.673 18.118 13.5546 18.31C13.3713 18.6072 13.3235 18.9684 13.4231 19.303C13.4876 19.5192 13.6485 19.7292 13.9703 20.1492C14.2921 20.5692 14.453 20.7793 14.645 20.8977C14.9422 21.081 15.3034 21.1288 15.638 21.0291C15.8542 20.9647 16.0642 20.8038 16.4842 20.482L18.2763 19.1099M18.2763 19.1099C18.6505 19.5984 19.3499 19.691 19.8383 19.3168L24.9191 15.4245M24.9191 15.4245C24.6535 15.0142 24.4993 14.5251 24.4993 14C24.4993 12.5503 25.6746 11.375 27.1243 11.375C27.8978 11.375 28.5931 11.7096 29.0735 12.2418'
+        stroke='#22202A'
+        stroke-width='1.5'
+      />
+      <path d='M11.375 26.25H18.375' stroke='#22202A' strokeWidth='1.5' strokeLinecap='round' />
+    </svg>
+  );
+}
+
+export default MicroscopeIcon();
